node_engine/requirements.txt
psutil
Flask
Flask-MQTT
Flask-SocketIO
docker==3.4.1
APScheduler
python-socketio[client]
GPUtil
tabulate
requests
gitpython
numpy
celery==4.4.2
redis
pyshark
shapely